+++ src/contracts/iauthentication.ts
