+++ src/contracts/ichoose-dialog.ts
