+++ src/contracts/idynamic-ui.ts
