+++ src/contracts/index.ts
@@ -1,6 +1,6 @@
-export * from './processengine';
-export * from './bpmnmodeler';
-export * from './dynamic-ui';
-export * from './choose-dialog';
-export * from './authentication';
-export * from './eventbus';
+export * from './iprocessengine';
+export * from './ibpmnmodeler';
+export * from './idynamic-ui';
+export * from './ichoose-dialog';
+export * from './iauthentication';
+export * from './ieventbus';
